+++ zeus-fe/App.tsx
@@ -1,17 +1,14 @@
 import { StatusBar } from 'expo-status-bar'
 import * as React from 'react'
-import { useEffect, useState } from 'react'
-import { Button, StyleSheet, Text, View, TextInput, Alert, Pressable} from 'react-native'
-import { Picker } from '@react-native-picker/picker'
+import { useState } from 'react'
+import { StyleSheet, Text, View } from 'react-native'
 import MapView from 'react-native-maps'
 import { useFonts, Poppins_700Bold } from '@expo-google-fonts/poppins'
-import { avoids, stations, metroExits, travelModes } from './Utils/MetroData'
+import { avoids, metroExits, travelModes, initialRegion } from './Constants'
 import MapViewDirections from 'react-native-maps-directions'
-import { StationPicker } from './Components/StationPicker'
 import { DestinationInput } from './Components/DestinationInput'
 // @ts-ignore
 import { GOOGLE_API_KEY, GOOGLE_API_BASE_URL } from '@env'
-import { initialRegion } from './Constants'
 
 export default function App() {
   useFonts({ Poppins_700Bold })
@@ -21,41 +18,16 @@ export default function App() {
   const [destination, setDestination] = useState({ latitude: 0, longitude: 0 })
   const [startLocation, setStartLocation] = useState({ latitude: 0, longitude: 0 })
   const [directions, setDirections] = useState('')
-  const [ stationsList, setStationsList ] = useState(stations)
-  const [ metroExitsList, setMetroExitisList ] = useState(metroExits)
-  const [ selectedStation, setSelectedStation ] = useState<string>('')
-  const [ exitsForSelectedStation, setExitsForSelectedStation ] = useState<MetroExit[]>()
-  //const [ urls, setUrls ] = useState<any[]>()
   const [error, setError] = useState('')
-
-  useEffect(() => {
-    //console.log(`stations is: ${JSON.stringify(stationsList)}`)
-    console.log(`selectedStation is: ${selectedStation}`)
-    console.log(`exitsForSelectedStation is: ${JSON.stringify(exitsForSelectedStation)}`)
-  },[selectedStation])
-
-  useEffect(() => {
-    //console.log(`directions: is: ${JSON.stringify(directions)}`)
-  },[directions])
-
-  const handleSelectedStation = (pickedStation: string) => {
-    //console.log(`pickedStation is: ${pickedStation}`)
-    setSelectedStation(pickedStation)
-    let tmp: Station = stationsList.filter( (it) => it.name.toLowerCase() === pickedStation.toLocaleLowerCase())[0]
-    //console.log(`tmp is: ${JSON.stringify(tmp)}`)
-    let stationExitsList = metroExitsList.filter( (it) => it.stationId === tmp.id )
-    //console.log(`stationExitsList is: ${JSON.stringify(stationExitsList)}`)
-    setExitsForSelectedStation(stationExitsList)
-
-  }
 
   const handleSubmit = async () => {
     let urls: string[] = []
-    exitsForSelectedStation?.forEach( (it) => {
-      urls = [...urls, `${GOOGLE_API_BASE_URL}origin=${it.latitude},${it.longitude}&destination=${selectedDestination}&mode=${travelModes[1].name}&avoid=${avoid}&key=${GOOGLE_API_KEY}`]
+    metroExits.forEach((it) => {
+      urls = [
+        ...urls,
+        `${GOOGLE_API_BASE_URL}origin=${it.latitude},${it.longitude}&destination=${selectedDestination}&mode=${travelModes[1].name}&avoid=${avoid}&key=${GOOGLE_API_KEY}`,
+      ]
     })
-    console.log(`urls is: ${urls}`)
-    
     try {
       const response = await Promise.all(
         urls.map(async (url) => {
@@ -63,7 +35,7 @@ export default function App() {
           return resp.json()
         }),
       )
-      console.log(`response: is: ${JSON.stringify(response)}`)
+      console.log(`response: is: ${JSON.stringify(response[0])}`)
       if (response[0].status === 'NOT_FOUND') return setError('Address not found! Please input a valid address.')
 
       let tempValue = 13600000
@@ -95,29 +67,23 @@ export default function App() {
   return (
     <View style={styles.container}>
       <Text style={styles.titleText}>Metro Exits</Text>
-      
-      <StationPicker selectedStation= {selectedStation} 
-      stationsList= {stationsList} 
-      handleSelectedStation={handleSelectedStation}
+      <DestinationInput
+        selectedDestination={selectedDestination}
+        setSelectedDestination={setSelectedDestination}
+        handleSubmit={handleSubmit}
       />
-
-      <DestinationInput selectedDestination={selectedDestination}
-      setSelectedDestination= {setSelectedDestination}
-      handleSubmit={handleSubmit}
-      />
-
       <MapView style={styles.map} initialRegion={initialRegion}>
         <MapViewDirections mode="WALKING" origin={startLocation} destination={destination} apikey={GOOGLE_API_KEY} />
       </MapView>
-      {
-        preferredExit ? 
+      {preferredExit ? (
         <>
           <View>
-            <Text style={{marginTop: 6, height: 30}}>Take the exit: {preferredExit}</Text>
+            <Text style={{ marginTop: 6, height: 30 }}>Take the exit: {preferredExit}</Text>
           </View>
         </>
-        : <></>
-      }
+      ) : (
+        <></>
+      )}
       <StatusBar style="auto" />
     </View>
   )
@@ -130,7 +96,7 @@ const styles = StyleSheet.create({
     backgroundColor: '#D9D9D9',
     alignItems: 'center',
     justifyContent: 'center',
-    marginTop: 20
+    marginTop: 20,
   },
   map: {
     width: '100%',
@@ -147,6 +113,6 @@ const styles = StyleSheet.create({
     fontWeight: '700',
     fontSize: 40,
     marginTop: 0,
-    paddingTop: 0
+    paddingTop: 0,
   },
 })
