--- zeus-fe/Utils/MetroData.ts
@@ -1,171 +1,0 @@
-export const stations = [
-  {
-    id: 0,
-    name: 'Hötorget',
-    imgSrc: '',
-  },
-  {
-    id: 1,
-    name: 'Rådmansgatan',
-    imgSrc: '',
-  },
-]
-
-export const travelModes = [
-  {
-    id: 0,
-    name: 'driving',
-  },
-  {
-    id: 1,
-    name: 'walking',
-  },
-  {
-    id: 2,
-    name: 'bicycling',
-  },
-  {
-    id: 3,
-    name: 'transit',
-  },
-]
-
-export const avoids = [
-  {
-    id: 0,
-    name: 'tolls',
-  },
-  {
-    id: 1,
-    name: 'highways',
-  },
-  {
-    id: 2,
-    name: 'ferries',
-  },
-]
-
-export const metroExits = [
-  {
-    id: 0,
-    name: 'Tunnelgatan 5A',
-    stationId: 0,
-    latitude: '59.336571',
-    longitude: '18.062832',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 2,
-    name: 'Olof palmes gata 7T',
-    stationId: 0,
-    latitude: '59.33633',
-    longitude: '18.062457',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 3,
-    name: 'Kungsgatan 36',
-    stationId: 0,
-    latitude: '59.335739',
-    longitude: '18.064088',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 4,
-    name: 'Kungsgatan 40',
-    stationId: 0,
-    latitude: '59.335488',
-    longitude: '18.062961',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 5,
-    name: 'Kungsgatan 39',
-    stationId: 0,
-    latitude: '59.335581',
-    longitude: '18.064356',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 6,
-    name: 'Kungsgatan 41',
-    stationId: 0,
-    latitude: '59.335307',
-    longitude: '18.063143',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 7,
-    name: 'Malmskillnadsgatan 27A',
-    stationId: 0,
-    latitude: '59.334536',
-    longitude: '18.065482',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 7,
-    name: 'Sveavägen 16',
-    stationId: 0,
-    latitude: '59.334393',
-    longitude: '18.064892',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 9,
-    name: 'Sveavägen 17A',
-    stationId: 0,
-    latitude: '59.334426',
-    longitude: '18.063755',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-
-  {
-    id: 10,
-    name: 'Rehnsgatan',
-    stationId: 1,
-    latitude: '59.342822',
-    longitude: '18.059576',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 11,
-    name: 'Sveavägen 98T',
-    stationId: 1,
-    latitude: '59.3423633',
-    longitude: '18.0552446',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 12,
-    name: 'Sveavägen 67B',
-    stationId: 1,
-    latitude: '59.3421262',
-    longitude: '18.0546131',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 13,
-    name: 'Sveavägen 55',
-    stationId: 1,
-    latitude: '59.3404744',
-    longitude: '18.056115',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 14,
-    name: 'Rådmansgatan södra',
-    stationId: 1,
-    latitude: '59.340275',
-    longitude: '18.058766',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-  {
-    id: 15,
-    name: 'Sveavägen 80A',
-    stationId: 1,
-    latitude: '59.3403637',
-    longitude: '18.0569798',
-    imgSrc: ['http://img1', 'http://img2', 'http://img3'],
-  },
-
-]
